cpython: drop bindgen c_int compatibility casts
User request: after switching bindgen macro constants to c_int-compatible defaults, remove the extra casts added for the previous unsigned define output.

Decision: simplify only the call sites and comparisons that were compensating for bindgen's old u32 macro defaults, while keeping unrelated c_ulong handling unchanged because type flags still cross an unsigned ABI boundary.

Agent-Prompt: 现在 bindgen 对 define 的类型默认是 c_int，导致需要 as libc::c_int (commit dd625af5d54ca655dd8377e7c23122a467f45304)，我想改成默认 c_int 看看
Agent-Model: gpt-5.4

+++ src/buffer.rs
@@ -167,11 +167,7 @@ impl PyBuffer {
             let mut buf = Box::new(mem::zeroed::<ffi::Py_buffer>());
             err::error_on_minusone(
                 py,
-                ffi::PyObject_GetBuffer(
-                    obj.as_ptr(),
-                    &mut *buf,
-                    ffi::PyBUF_FULL_RO as libc::c_int,
-                ),
+                ffi::PyObject_GetBuffer(obj.as_ptr(), &mut *buf, ffi::PyBUF_FULL_RO),
             )?;
             validate(&buf);
             Ok(PyBuffer(buf))
